Release-it | Reorganize files
